requirements.txt
# Core
aiogram==3.15.0
python-dotenv==1.0.1

# LLM
groq==0.10.0

# Voice Processing
SpeechRecognition==3.10.4
pydub==0.25.1

# HTTP Client
aiohttp==3.10.5
httpx==0.27.0

# Database
supabase==2.7.1
postgrest==2.0.0
realtime==0.2.0

# Image Generation
google-generativeai==0.3.2
replicate==0.25.0
Pillow==10.2.0
aiofiles==23.2.1

# Image Processing
python-multipart==0.0.9
imageio==2.34.0
pybase64==1.3.2

# Utilities
redis==5.0.1  # Для кэширования (опционально)
python-json-logger==2.0.7
schedule==1.2.1
watchdog==3.0.0

# Development
black==24.1.1
isort==5.13.2
pytest==7.4.4
pytest-asyncio==0.23.2